+++ test/libsolidity/syntaxTests/array/slice/assign_to_storage.sol
@@ -1,0 +1,8 @@
+contract c {
+  bytes public b;
+  function f() public {
+    b = msg.data[:];
+  }
+}
+// ----
+// TypeError: (63-74): Type bytes calldata slice is not implicitly convertible to expected type bytes storage ref.